api/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
SQLAlchemy==2.0.29
python-dotenv==1.0.1
APScheduler==3.10.4
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
psutil==5.9.8
boto3==1.34.83
google-cloud-monitoring==2.21.0
